pyproject.toml
[project]
name = "gut-brain-ibs-pipeline"
version = "0.1.0"
description = "Windows-first reproducible Snakemake pipeline for gut-brain IBS genetics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "pyarrow>=15.0",
  "duckdb>=1.1",
  "pyyaml>=6.0",
  "requests>=2.32",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "upsetplot>=0.9",
  "adjustText>=1.3",
  "matplotlib-venn>=1.1",
  "tabulate>=0.9",
  "h5py>=3.10",
  "anndata>=0.10",
  "scipy>=1.12",
  "psutil>=5.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "snakemake>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
